move `network::messages::MessageContainer` to `network::MessageContainer`
